TST: Use global precision constants in tests

+++ tests/operators/test_propagation.py
@@ -5,6 +5,7 @@ import unittest
 
 import numpy as np
 from tike.operators import Propagation
+import tike.precision
 
 from .util import random_complex, OperatorTests
 
@@ -26,13 +27,11 @@ class TestPropagation(unittest.TestCase, OperatorTests):
         self.operator.__enter__()
         self.xp = self.operator.xp
         np.random.seed(0)
-        self.m = self.xp.asarray(random_complex(nwaves, probe_shape,
-                                                probe_shape),
-                                 dtype='complex64')
+        self.m = self.xp.asarray(
+            random_complex(nwaves, probe_shape, probe_shape))
         self.m_name = 'nearplane'
-        self.d = self.xp.asarray(random_complex(nwaves, probe_shape,
-                                                probe_shape),
-                                 dtype='complex64')
+        self.d = self.xp.asarray(
+            random_complex(nwaves, probe_shape, probe_shape))
         self.d_name = 'farplane'
         self.kwargs = {}
         print(self.operator)
